20131122-mw-builder.utf8

+++ mw-builder/skin/board/mw.basic/write_update.skin.php
@@ -182,8 +182,17 @@ if ($mw_basic[cf_change_image_size] && $member[mb_level] >= $mw_basic[cf_change_
 }
 else
 {*/
-    $thumb_file = "$thumb_path/{$wr_id}";
-    unlink($thumb_file);
+    $thumb_file = mw_thumb_jpg("$thumb_path/{$wr_id}");
+    $thumb2_file = mw_thumb_jpg("$thumb2_path/{$wr_id}");
+    $thumb3_file = mw_thumb_jpg("$thumb3_path/{$wr_id}");
+    $thumb4_file = mw_thumb_jpg("$thumb4_path/{$wr_id}");
+    $thumb5_file = mw_thumb_jpg("$thumb5_path/{$wr_id}");
+
+    if (file_exists($thumb_file)) unlink($thumb_file);
+    if (file_exists($thumb2_file)) unlink($thumb2_file);
+    if (file_exists($thumb3_file)) unlink($thumb3_file);
+    if (file_exists($thumb4_file)) unlink($thumb4_file);
+    if (file_exists($thumb5_file)) unlink($thumb5_file);
 
     $thumb_file = "";
     $file = mw_get_first_file($bo_table, $wr_id, true);
@@ -193,42 +202,56 @@ else
         mw_make_thumbnail($mw_basic[cf_thumb_width], $mw_basic[cf_thumb_height], $source_file, $thumb_file, $mw_basic[cf_thumb_keep]);
         if ($mw_basic[cf_thumb2_width])
             @mw_make_thumbnail($mw_basic[cf_thumb2_width], $mw_basic[cf_thumb2_height], $source_file,
-                "{$thumb2_path}/{$wr_id}", $mw_basic[cf_thumb2_keep]);
+                "{$thumb2_file}", $mw_basic[cf_thumb2_keep]);
         if ($mw_basic[cf_thumb3_width]) {
             @mw_make_thumbnail($mw_basic[cf_thumb3_width], $mw_basic[cf_thumb3_height], $source_file,
-                "{$thumb3_path}/{$wr_id}", $mw_basic[cf_thumb3_keep]);
+                "{$thumb3_file}", $mw_basic[cf_thumb3_keep]);
         }
         if ($mw_basic[cf_thumb4_width]) {
             @mw_make_thumbnail($mw_basic[cf_thumb4_width], $mw_basic[cf_thumb4_height], $source_file,
-                "{$thumb4_path}/{$wr_id}", $mw_basic[cf_thumb4_keep]);
+                "{$thumb4_file}", $mw_basic[cf_thumb4_keep]);
         }
         if ($mw_basic[cf_thumb5_width]) {
             @mw_make_thumbnail($mw_basic[cf_thumb5_width], $mw_basic[cf_thumb5_height], $source_file,
-                "{$thumb5_path}/{$wr_id}", $mw_basic[cf_thumb5_keep]);
+                "{$thumb5_file}", $mw_basic[cf_thumb5_keep]);
         }
     } else {
+        $is_thumb = false;
         $thumb_file = "$thumb_path/{$wr_id}";
-        preg_match("/<img.*src=\\\"(.*)\\\"/iUs", stripslashes($wr_content), $match);
-        if ($match[1]) {
-            //$match[1] = str_replace($g4[url], "..", $match[1]);
-            $match[1] = preg_replace("/(http:\/\/.*)\/data\//i", "../data/", $match[1]);
-            if (file_exists($match[1])) {
-                mw_make_thumbnail($mw_basic[cf_thumb_width], $mw_basic[cf_thumb_height], $match[1], $thumb_file, $mw_basic[cf_thumb_keep]);
-                if ($mw_basic[cf_thumb2_width])
-                    @mw_make_thumbnail($mw_basic[cf_thumb2_width], $mw_basic[cf_thumb2_height], $match[1],
-                        "{$thumb2_path}/{$wr_id}", $mw_basic[cf_thumb2_keep]);
-                if ($mw_basic[cf_thumb3_width])
-                    @mw_make_thumbnail($mw_basic[cf_thumb3_width], $mw_basic[cf_thumb3_height], $match[1],
-                        "{$thumb3_path}/{$wr_id}", $mw_basic[cf_thumb3_keep]);
-                if ($mw_basic[cf_thumb4_width])
-                    @mw_make_thumbnail($mw_basic[cf_thumb4_width], $mw_basic[cf_thumb4_height], $match[1],
-                        "{$thumb4_path}/{$wr_id}", $mw_basic[cf_thumb4_keep]);
-                if ($mw_basic[cf_thumb5_width])
-                    @mw_make_thumbnail($mw_basic[cf_thumb5_width], $mw_basic[cf_thumb5_height], $match[1],
-                        "{$thumb5_path}/{$wr_id}", $mw_basic[cf_thumb5_keep]);
-            }
-        } else {
-            @unlink($thumb_file);
+        preg_match_all("/<img.*src=\\\"(.*)\\\"/iUs", stripslashes($wr_content), $matchs);
+        for ($i=0, $m=count($matchs[1]); $i<$m; ++$i) {
+            $mat = $matchs[1][$i];
+            if (strstr($mat, "mw.basic.comment.image")) $mat = '';
+            if (strstr($mat, "mw.emoticon")) $mat = '';
+            if (preg_match("/cheditor[0-9]\/icon/i", $mat)) $mat = '';
+            if ($mat) {
+                //$mat = str_replace($g4[url], "..", $mat);
+                $mat = preg_replace("/(http:\/\/.*)\/data\//i", "../data/", $mat);
+                if (file_exists($mat)) {
+                    mw_make_thumbnail($mw_basic[cf_thumb_width], $mw_basic[cf_thumb_height], $mat, $thumb_file, $mw_basic[cf_thumb_keep]);
+                    if ($mw_basic[cf_thumb2_width])
+                        @mw_make_thumbnail($mw_basic[cf_thumb2_width], $mw_basic[cf_thumb2_height], $mat,
+                            "{$thumb2_file}", $mw_basic[cf_thumb2_keep]);
+                    if ($mw_basic[cf_thumb3_width])
+                        @mw_make_thumbnail($mw_basic[cf_thumb3_width], $mw_basic[cf_thumb3_height], $mat,
+                            "{$thumb3_file}", $mw_basic[cf_thumb3_keep]);
+                    if ($mw_basic[cf_thumb4_width])
+                        @mw_make_thumbnail($mw_basic[cf_thumb4_width], $mw_basic[cf_thumb4_height], $mat,
+                            "{$thumb4_file}", $mw_basic[cf_thumb4_keep]);
+                    if ($mw_basic[cf_thumb5_width])
+                        @mw_make_thumbnail($mw_basic[cf_thumb5_width], $mw_basic[cf_thumb5_height], $mat,
+                            "{$thumb5_file}", $mw_basic[cf_thumb5_keep]);
+                    $is_thumb = true;
+                    break;
+                }
+            } // if 
+        } // for 
+        if (!$is_thumb) {
+            if (file_exists($thumb_file)) unlink($thumb_file);
+            if (file_exists($thumb2_file)) unlink($thumb2_file);
+            if (file_exists($thumb3_file)) unlink($thumb3_file);
+            if (file_exists($thumb4_file)) unlink($thumb4_file);
+            if (file_exists($thumb5_file)) unlink($thumb5_file);
         }
     }
 //}
@@ -497,6 +520,9 @@ if ($mw_basic[cf_social_commerce]) include("$social_commerce_path/write_update.s
 
 // 재능마켓
 if ($mw_basic[cf_talent_market]) include("$talent_market_path/write_update.skin.php");
+
+// 마케팅DB
+if ($mw_basic[cf_marketdb]) include("$marketdb_path/write_update.skin.php");
 
 // 구글지도
 if ($mw_basic[cf_google_map]) {
